[FEATURE] Further tweaks to 2019 design
I wanted to try and make the checker fit in more with the look and feel of the main site.

Main changes I've made are to use a monospace font, darken the white areas of the site and change links from purple, to blue + purple hover. I also removed any rounding of buttons etc so that everything returned to being square edged.

I also removed the author part of the pull request info, I figured if someone was checkign their results the author part is a bit redundant so for brevity sake removed it.

Also fixed a bug where 'checker' was causing unnceccesary horizontal scroll.